backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
pillow>=10.0.0
pyyaml==6.0.1
ultralytics==8.0.227
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
websockets==12.0
python-dotenv==1.0.0
sendgrid==6.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
